% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Estimator}
\usage{
bootstrap(n=200, x, y, plot=FALSE)
}
\arguments{
\item{n}{is a positive integer that specifies how many bootstrap replicates should be used for estimation.}

\item{x}{is a vector of numeric values to be used with y to estimate correlation and bootstrap estimates.}

\item{y}{is a vector of numeric values to be used with x to estimate correlation and bootstrap estimates. y has to be of the same length as x.}

\item{plot}{is a boolean value that specifies whether a plot should be created or not}
}
\value{
returns the Correlation, as well as bootstrap estimation of correlation(BootstrapCor), standard error(StandardError), bias(Bias) and confidence interval(CI).
}
\description{
This function takes 3 arguments to estimate the correlation between them, as well as the bootstrap estimation of the correlation, standard error, bias and confidence interval, using these arguments.
}
\examples{
#Add examples for the function to explain to others how they can be used
bootstrap(n=1000, x=rnorm(30),y=rnorm(30), plot = TRUE)
}
\author{
Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
}
\keyword{bias}
\keyword{bootstrap}
\keyword{confidence}
\keyword{correlation}
\keyword{error}
\keyword{estimation}
\keyword{interval}
\keyword{standard}

