% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.gof.R
\name{chi.gof}
\alias{chi.gof}
\title{Chi-squared Goodness of Fit test}
\usage{
chi.gof(x,p)
}
\arguments{
\item{x}{is a categorical variable vector of numbers, counting number of each incident.}

\item{p}{is an optional vector of probabilities. If unused, replaced by a uniformly distributed vector of probabilities.}
}
\value{
returns the Chi-squared goodness of fit value, calculated using Pearson's Chi-squared test
}
\description{
This function takes 1 or 2 arguments to calculate the chi squared goodness of fit value from these arguments.
}
\examples{
#Add examples for the function to explain to others how they can be used
chi.gof(c(1,2,3),c(0.25,0.6,0.15))
}
\author{
Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
}
\keyword{chi}
\keyword{distribution}
\keyword{fit}
\keyword{goodness}
\keyword{pearson}
\keyword{squared}
\keyword{value}

