% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.probability.R
\name{chi.probability}
\alias{chi.probability}
\title{Chi Probability}
\usage{
chi.probability(x,df = 1,n = 100)
}
\arguments{
\item{x}{specifies the value that should be evaluted by chi.probability}

\item{df}{is a positive integer that specifies the degrees of freedom in the chi-squared distribution.}

\item{n}{is a positive integer that specifies how many random variables that is used for the evaluation.}
}
\value{
returns the probability that the chi-squared distribution is larger than the value of x.
}
\description{
This function takes 3 arguments to estimate the probability that the chi-squared distribution is larger than an input value x.
}
\examples{
#Add examples for the function to explain to others how they can be used
chi.probability(5,1,100)
}
\author{
Emil H. Andersen \cr
Department of Mathematics and Computer Science (IMADA) \cr
University of Southern Denmark, Denmark \cr
\email{emila14@student.sdu.dk} \cr
}
\keyword{carlo}
\keyword{chi}
\keyword{distribution}
\keyword{estimation}
\keyword{matrix}
\keyword{monte}
\keyword{simulation}

